lttng: Do not rebuild the SS in every single state system test
At some point we changed the state system tests to rebuild the whole
thing in the @Before, because at that point AssumeTrue() inside a
@BeforeClass did not work very well. Unfortunately this made running
the tests much longer.

Instead, really build the analysis/state system once per test class,
but only keep the AssumeTrue() in the main @Before that is ran every
test. That check will test a static boolean set by the @BeforeClass.

tl;dr Make the state system tests ~20 times faster.

Reviewed-on: https://git.eclipse.org/r/36184
Tested-by: Hudson CI
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>

+++ org.eclipse.tracecompass.lttng2.kernel.core.tests/src/org/eclipse/tracecompass/lttng2/kernel/core/tests/stateprovider/PartialStateSystemTest.java
@@ -21,7 +21,6 @@ import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.tracecompass.internal.lttng2.kernel.core.stateprovider.LttngKernelStateProvider;
-import org.eclipse.tracecompass.statesystem.core.ITmfStateSystem;
 import org.eclipse.tracecompass.statesystem.core.exceptions.TimeRangeException;
 import org.eclipse.tracecompass.tmf.core.exceptions.TmfAnalysisException;
 import org.eclipse.tracecompass.tmf.core.statesystem.ITmfStateProvider;
@@ -29,7 +28,8 @@ import org.eclipse.tracecompass.tmf.core.statesystem.TmfStateSystemAnalysisModul
 import org.eclipse.tracecompass.tmf.core.trace.ITmfTrace;
 import org.eclipse.tracecompass.tmf.core.trace.TmfTraceManager;
 import org.eclipse.tracecompass.tmf.ctf.core.trace.CtfTmfTrace;
-import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Test;
 
 /**
@@ -41,11 +41,20 @@ public class PartialStateSystemTest extends StateSystemTest {
 
     private static final @NonNull String TEST_FILE_NAME = "test-partial";
 
-    private File stateFile;
-    private TestLttngKernelAnalysisModule module;
+    private static File stateFile;
+    private static TestLttngKernelAnalysisModule module;
 
-    @Override
-    protected ITmfStateSystem initialize() {
+    /**
+     * Test class setup
+     */
+    @BeforeClass
+    public static void initialize() {
+        if (!testTrace.exists()) {
+            traceIsPresent = false;
+            return;
+        }
+        traceIsPresent = true;
+
         stateFile = new File(TmfTraceManager.getSupplementaryFileDir(testTrace.getTrace()) + TEST_FILE_NAME);
         if (stateFile.exists()) {
             stateFile.delete();
@@ -59,20 +68,26 @@ public class PartialStateSystemTest extends StateSystemTest {
         }
         module.schedule();
         assertTrue(module.waitForCompletion());
-        return module.getStateSystem();
+
+        fixture = module.getStateSystem();
     }
 
     /**
      * Class clean-up
      */
-    @After
-    public void cleanup() {
+    @AfterClass
+    public static void cleanup() {
         if (module != null) {
             module.dispose();
         }
         if (stateFile != null) {
             stateFile.delete();
         }
+        if (fixture != null) {
+            fixture.dispose();
+        }
+        module = null;
+        fixture = null;
     }
 
     /**
